server/python_ai/requirements.txt
# ============================================
# REQUIREMENTS.TXT - Python AI Backend (Modular)
# ============================================
# Dự án: AI English Learning App - IELTS Writing & Speaking Feedback
# Backend: Flask + Google Gemini AI + Whisper + TTS + NLP
# ============================================


# --- WEB FRAMEWORK & SERVER ---
Flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0

# --- DATABASE ---
PyMongo>=4.6.0
motor>=3.3.2

# --- AI & NLP CORE ---
google-genai>=0.3.0
spacy>=3.7.0
language-tool-python>=2.8.0
pyspellchecker>=0.8.1
textstat>=0.7.3

# --- SPEECH RECOGNITION & TTS ---
openai-whisper>=20231117
faster-whisper>=1.0.3
edge-tts>=6.1.10


# --- AUDIO/VIDEO PROCESSING ---
ffmpeg-python>=0.2.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0

# --- DATA SCIENCE & ML ---
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0

# --- HTTP & API ---
requests>=2.31.0
python-multipart>=0.0.6

# --- UTILITIES ---
python-dateutil>=2.8.2
pysbd>=0.3.4
lexical-diversity>=0.1.1
transformers>=4.35.0
torch>=2.1.0
onnxruntime>=1.16.0
optimum[onnxruntime]>=1.15.0

# ============================================
# INSTALLATION NOTES:
# ============================================
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate virtual environment:
#    Windows: venv\Scripts\activate
#    Linux/Mac: source venv/bin/activate
#
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# 4. Download spaCy model:
#    python -m spacy download en_core_web_md
#
# 5. Install FFmpeg (required for audio processing):
#    Windows: Download from ffmpeg.org
#    Linux: sudo apt-get install ffmpeg
#    Mac: brew install ffmpeg
#
# 6. Set environment variables in .env:
#    GEMINI_API_KEY=your_api_key_here
# ============================================
